Refactor: DatasetWriteUseCase to reuse logic shared across different use cases
