+++ src/core/domain/useCases/UseCase.ts
@@ -1,4 +1,3 @@
 export interface UseCase<T> {
-  // eslint-disable-next-line  @typescript-eslint/no-explicit-any
-  execute(...args: any[]): Promise<T>
+  execute(...args: unknown[]): Promise<T>
 }
