+++ src/datasets/domain/useCases/CreateDataset.ts
@@ -1,25 +1,18 @@
-import { UseCase } from '../../../core/domain/useCases/UseCase'
 import { IDatasetsRepository } from '../repositories/IDatasetsRepository'
-import { NewDatasetDTO, NewDatasetMetadataBlockValuesDTO } from '../dtos/NewDatasetDTO'
+import { NewDatasetDTO } from '../dtos/NewDatasetDTO'
 import { NewResourceValidator } from '../../../core/domain/useCases/validators/NewResourceValidator'
 import { IMetadataBlocksRepository } from '../../../metadataBlocks/domain/repositories/IMetadataBlocksRepository'
-import { MetadataBlock } from '../../../metadataBlocks'
 import { CreatedDatasetIdentifiers } from '../models/CreatedDatasetIdentifiers'
 import { ROOT_COLLECTION_ALIAS } from '../../../collections/domain/models/Collection'
+import { DatasetWriteUseCase } from './DatasetWriteUseCase'
 
-export class CreateDataset implements UseCase<CreatedDatasetIdentifiers> {
-  private datasetsRepository: IDatasetsRepository
-  private metadataBlocksRepository: IMetadataBlocksRepository
-  private newDatasetValidator: NewResourceValidator
-
+export class CreateDataset extends DatasetWriteUseCase<CreatedDatasetIdentifiers> {
   constructor(
     datasetsRepository: IDatasetsRepository,
     metadataBlocksRepository: IMetadataBlocksRepository,
     newDatasetValidator: NewResourceValidator
   ) {
-    this.datasetsRepository = datasetsRepository
-    this.metadataBlocksRepository = metadataBlocksRepository
-    this.newDatasetValidator = newDatasetValidator
+    super(datasetsRepository, metadataBlocksRepository, newDatasetValidator)
   }
 
   /**
@@ -37,23 +30,7 @@ export class CreateDataset implements UseCase<CreatedDatasetIdentifiers> {
     collectionId = ROOT_COLLECTION_ALIAS
   ): Promise<CreatedDatasetIdentifiers> {
     const metadataBlocks = await this.getNewDatasetMetadataBlocks(newDataset)
-
-    this.newDatasetValidator.validate(newDataset, metadataBlocks)
-
-    return this.datasetsRepository.createDataset(newDataset, metadataBlocks, collectionId)
-  }
-
-  async getNewDatasetMetadataBlocks(newDataset: NewDatasetDTO): Promise<MetadataBlock[]> {
-    const metadataBlocks: MetadataBlock[] = []
-    await Promise.all(
-      newDataset.metadataBlockValues.map(
-        async (metadataBlockValue: NewDatasetMetadataBlockValuesDTO) => {
-          metadataBlocks.push(
-            await this.metadataBlocksRepository.getMetadataBlockByName(metadataBlockValue.name)
-          )
-        }
-      )
-    )
-    return metadataBlocks
+    this.getNewDatasetValidator().validate(newDataset, metadataBlocks)
+    return this.getDatasetsRepository().createDataset(newDataset, metadataBlocks, collectionId)
   }
 }
